♻️ [refactor] 로직 분리해 useCamera, useGallery 훅 생성
